Refactor package location, fix getInitials
