+++ app/src/main/java/com/example/tasky/agenda/data/mapper/TaskMapper.kt
@@ -2,7 +2,7 @@ package com.example.tasky.agenda.data.mapper
 
 import com.example.tasky.agenda.data.local.model.TaskEntity
 import com.example.tasky.agenda.data.remote.model.TaskResponse
-import com.example.tasky.agenda.domain.model.AgendaItem
+import com.example.tasky.core.domain.model.AgendaItem
 
 fun TaskResponse.toDomain() =
     AgendaItem.Task(
