+++ app/src/main/java/com/example/tasky/itemdetail/presentation/components/VisitorsTypePill.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.tasky.itemdetail.domain.VisitorType
+import com.example.tasky.itemdetail.presentation.model.VisitorType
 import com.example.tasky.ui.theme.Black
 import com.example.tasky.ui.theme.DarkGray
 import com.example.tasky.ui.theme.Light2
